==> Codes to extract data/0. Prepare ISIN for WRDS.ipynb <==
import pandas as pd
import numpy as np

Latin_companies = pd.read_excel('2.16 S&P Latin America companies.xlsx')

Latin_companies.head()

np.sum(Latin_companies.isnull(), axis = 0)

Latin_companies = Latin_companies[Latin_companies['SP_EXCHANGE'].notnull()]
Latin_companies = Latin_companies[Latin_companies['SP_GEOGRAPHY'].notnull()]
np.sum(Latin_companies.isnull(), axis = 0)

isin_list = Latin_companies['SP_ISIN'].to_list()

print(len(isin_list))
isin_list[:5]

with open('temp.txt', 'w') as file:
  for i in isin_list:
    file.write(str(i) + '\n')
