==> Codes to extract data/1. Data cleaning 2.18 - with WRDS balance sheet dataset.ipynb <==
import pandas as pd
import numpy as np

def check(df):
  num_company = df['isin'].nunique()
  num_rows = df.shape[0]
  print(f'Number of companies: {num_company}')
  print(f'Number of rows: {num_rows}')

def get_num_finance(df):
  return df[['isin', 'is_finance']].drop_duplicates()['is_finance'].sum()

# Check the WRDS balance sheet data

balance_sheet = pd.read_csv('2.16 balance sheet.csv')

balance_sheet.head()

check(balance_sheet)

np.sort(balance_sheet['fyearq'].unique())

# remove the 2005 and 2024 data
balance_sheet_2006 = balance_sheet[~balance_sheet['fyearq'].isin([2005, 2024])]

np.sort(balance_sheet_2006['fyearq'].unique())

check(balance_sheet_2006)

# Deal with missing values in `balance_sheet_2006`

np.sum(balance_sheet_2006.isnull(), axis = 0)

# remove actq, ancq, lltq
balance_sheet_2006.drop(columns = ['actq', 'ancq', 'lltq'], inplace = True)

balance_sheet_2006.columns

balance_sheet_2006.columns = ['key', 'fiscal_quarter', 'fiscal_year', 'datadate', 'total_assets', 'total_liabilities', 'isin', 'company_name', 'gic_industries', 'iso_country_code']

balance_sheet_2006.head(1)

balance_sheet_2006['gic_industries'] = balance_sheet_2006['gic_industries'].astype(str)

# Change industry name

gic_codes_table = pd.read_excel('gic industry codes.xlsx')

gic_codes_table.head()

gic_codes_table = gic_codes_table[['Industry', 'Industry2']]
gic_codes_table.columns = ['industry_code', 'industry_name']
gic_codes_table.drop_duplicates(inplace = True)
gic_codes_table['industry_code'] = gic_codes_table['industry_code'].astype(str)
gic_codes_table.set_index('industry_code', inplace = True)
gic_codes_table.head()

def get_industry_name(df):
  code = df['gic_industries']
  series = gic_codes_table['industry_name']
  return series[code]

balance_sheet_2006['industry_name'] = balance_sheet_2006.apply(get_industry_name, axis = 1)

balance_sheet_2006.head(1)

balance_sheet_2006['is_finance'] = balance_sheet_2006['gic_industries'].str.startswith('40')
get_num_finance(balance_sheet_2006)

balance_sheet_2006[balance_sheet_2006['is_finance'] == True]['industry_name'].unique()

# Check the industries

balance_sheet_2006[['isin', 'industry_name']].drop_duplicates()['industry_name'].value_counts()

np.sum(balance_sheet_2006.isnull(), axis = 0)

# Remove the companies with no industry data

balance_sheet_2006 = balance_sheet_2006[balance_sheet_2006['industry_name'].notnull()]
check(balance_sheet_2006)

np.sum(balance_sheet_2006.isnull(), axis = 0)

get_num_finance(balance_sheet_2006)

## Check the starting date of each company

balance_sheet_2006['datadate'] = pd.to_datetime(balance_sheet_2006['datadate'])
balance_sheet_2006['MM-DD'] = balance_sheet_2006['datadate'].dt.strftime('%m-%d')

balance_sheet_2006['MM-DD'].value_counts()

balance_sheet_2006['MM-DD'].value_counts().index

### Decision:
Drop unsuitable rows.

balance_sheet_2006 = balance_sheet_2006[~(balance_sheet_2006['MM-DD'].isin(['08-31', '05-31', '11-30', '10-31', '02-28', '07-31', '01-31', '04-30', '02-29']))]

check(balance_sheet_2006)

balance_sheet_2006['MM-DD'].value_counts()

# Change date columns + interpolation

quarterly_dates = pd.date_range(start='2006-01-01', end='2023-12-31', freq='Q').to_list()
correct_datetime = pd.DataFrame({'datadate': quarterly_dates})
correct_datetime.head()

# find the expected length of the dataframe
expected_length = len(quarterly_dates*(balance_sheet_2006['isin'].nunique()))
expected_length

correct_datetime.set_index('datadate', inplace = True)

company_list = list(balance_sheet_2006['isin'].unique())

check(balance_sheet_2006)

np.sum(balance_sheet_2006.isnull(), axis = 0)

first_company_info = balance_sheet_2006[balance_sheet_2006['isin'] == company_list[0]].set_index('datadate')
balance_sheet_adjusted = first_company_info.join(correct_datetime, how = 'outer').reset_index()

for column in ['key', 'isin', 'company_name', 'gic_industries', 'iso_country_code']:
    balance_sheet_adjusted[column] = balance_sheet_adjusted[balance_sheet_adjusted[column].notnull()][column].unique().item()

balance_sheet_adjusted['fiscal_quarter']=balance_sheet_adjusted['datadate'].dt.quarter
balance_sheet_adjusted['fiscal_year']=balance_sheet_adjusted['datadate'].dt.year

# interpolation:
balance_sheet_adjusted.set_index('datadate', inplace = True)
balance_sheet_adjusted['total_assets'] = balance_sheet_adjusted['total_assets'].interpolate(method='time', limit_direction='both')
balance_sheet_adjusted['total_liabilities'] = balance_sheet_adjusted['total_liabilities'].interpolate(method='time', limit_direction='both')
balance_sheet_adjusted.shape
balance_sheet_adjusted.reset_index(drop = False, inplace = True)

questionable_isin_dict = {}

for isin in company_list[1:]:
    temp = balance_sheet_2006[balance_sheet_2006['isin'] == isin].set_index('datadate')
    temp = temp.join(correct_datetime, how = 'outer').reset_index()
    if temp.shape[0] != 72:
      print(isin, list(temp[temp['datadate'].duplicated()]['datadate']))
      questionable_isin_dict = {**questionable_isin_dict, **{isin: list(temp[temp['datadate'].duplicated()]['datadate'])}}

    for column in ['key', 'isin', 'company_name', 'gic_industries', 'iso_country_code', 'industry_name', 'is_finance']:
      temp[column] = temp[temp[column].notnull()][column].unique().item()

    temp['fiscal_quarter']=temp['datadate'].dt.quarter
    temp['fiscal_year']=temp['datadate'].dt.year

    # interpolation:
    temp.set_index('datadate', inplace = True)
    temp['total_assets'] = temp['total_assets'].interpolate(method='time', limit_direction='both')
    temp['total_liabilities'] = temp['total_liabilities'].interpolate(method='time', limit_direction='both')
    temp.reset_index(drop = False, inplace = True)

    balance_sheet_adjusted = pd.concat([balance_sheet_adjusted, temp], axis = 0)

print(balance_sheet_adjusted.shape[0])

balance_sheet_adjusted['index'] = balance_sheet_adjusted.reset_index().index
for isin, time_list in questionable_isin_dict.items():
  temp = balance_sheet_adjusted[balance_sheet_adjusted['isin'] == isin]
  temp.head()
  print(temp[temp['datadate'].isin(time_list)].to_string())
  print('--------------')

check(balance_sheet_adjusted)

expected_length

balance_sheet_adjusted.drop(columns = ['MM-DD'], inplace = True)

balance_sheet_adjusted.drop_duplicates(subset = ['isin', 'datadate'], keep = 'last', inplace = True)
check(balance_sheet_adjusted)

balance_sheet_adjusted.drop(columns = ['index'], inplace = True)

get_num_finance(balance_sheet_adjusted)

# Check the rest of null values

check(balance_sheet_adjusted)

np.sum(balance_sheet_adjusted.isnull(), axis = 0)

balance_sheet_adjusted['industry_name'] = balance_sheet_adjusted.apply(get_industry_name, axis = 1)
balance_sheet_adjusted['is_finance'] = balance_sheet_adjusted['gic_industries'].str.startswith('40')

np.sum(balance_sheet_adjusted.isnull(), axis = 0)

check(balance_sheet_adjusted)

balance_sheet_adjusted.head(1)

get_num_finance(balance_sheet_adjusted)

# Export the interpolated balance sheet

#balance_sheet_adjusted.to_csv('2.18 interpolated balance sheet.csv', index=False)

# Export the isin list that will be used for stock data

isin_list = list(balance_sheet_adjusted['isin'].unique())
with open('2.18 isin list.txt', 'w') as f:
    for item in isin_list:
        f.write(f"{item}\n")
