==> Codes to extract data/3. Data Cleaning - stock price data 2.18.ipynb <==
import pandas as pd
import numpy as np

def check_null(df):
  return np.sum(df.isnull(), axis = 0)
def check_zero(df):
  return np.sum(df == 0, axis = 0)

stock_price = pd.read_csv('2.18 stock price.csv')

stock_price.head()

stock_price.shape

check_null(stock_price)

check_zero(stock_price)

stock_price[stock_price['Open']==0].head()

for column in ['Open', 'High', 'Low', 'Close', 'Adj Close', 'Volume']:
  stock_price[column] = np.where(stock_price[column] == 0, np.nan, stock_price[column])

check_null(stock_price)

check_null(stock_price) + 55837

stock_price['Date'] = pd.to_datetime(stock_price['Date'])

stock_price['Year'] = stock_price['Date'].dt.strftime('%Y')

stock_price = stock_price[stock_price['Year'] != '2024']

stock_price.shape[0]

isin_2006 = list(stock_price[stock_price['Year'] == '2006']['ISIN'].unique())
len(isin_2006)

stock_price_2006 = stock_price[stock_price['ISIN'].isin(isin_2006)]

stock_price_2006.shape[0]

check_null(stock_price_2006)

# Try to get all the dates

correct_datetime = pd.DataFrame({'datadate': list(np.sort(stock_price['Date'].unique()))})

correct_datetime.head()

correct_datetime.shape[0]

len(isin_2006)

# find the expected length of the dataframe
expected_length = correct_datetime.shape[0] * len(isin_2006)
expected_length

1253565 - 1197728 = 55837 rows interpolated.

# Time interpolation

stock_price_2006.head(1)

correct_datetime.set_index('datadate', inplace = True)

first_company_info = stock_price_2006[stock_price_2006['ISIN'] == isin_2006[0]].set_index('Date')
stock_price_adjusted = first_company_info.join(correct_datetime, how = 'outer').reset_index()

for column in ['Ticker', 'Exchange', 'ISIN']:
    stock_price_adjusted[column] = stock_price_adjusted[stock_price_adjusted[column].notnull()][column].unique().item()


# interpolation:
stock_price_adjusted.set_index('index', inplace = True)
for column in ['Open', 'High', 'Low', 'Close', 'Adj Close', 'Volume']:
  stock_price_adjusted[column] = stock_price_adjusted[column].interpolate(method='time', limit_direction='both')
stock_price_adjusted.reset_index(drop = False, inplace = True)

stock_price_adjusted.shape[0]

for isin in isin_2006[1:]:
    temp = stock_price_2006[stock_price_2006['ISIN'] == isin].set_index('Date')
    temp = temp.join(correct_datetime, how = 'outer').reset_index()

    for column in ['Ticker', 'Exchange', 'ISIN']:
      temp[column] = temp[temp[column].notnull()][column].unique().item()


    # interpolation:
    temp.set_index('index', inplace = True)
    for column in ['Open', 'High', 'Low', 'Close', 'Adj Close', 'Volume']:
      temp[column] = temp[column].interpolate(method='time', limit_direction='both')
    temp.reset_index(drop = False, inplace = True)

    stock_price_adjusted = pd.concat([stock_price_adjusted, temp], axis = 0)

print(stock_price_adjusted.shape[0])

stock_price_adjusted.columns = ['Date', 'Open', 'High', 'Low', 'Close', 'Adj Close', 'Volume', 'Ticker', 'Exchange', 'ISIN', 'Year']

stock_price_adjusted['Year'] = stock_price_adjusted['Date'].dt.strftime('%Y')

check_null(stock_price_adjusted)

# All prices to USD

stock_price_adjusted.head(1)

stock_price_adjusted.shape

## Load exchange rate dataset

exchange_rate = pd.read_excel('exchange rates.xlsx')

exchange_rate.head(2)

def get_exchange_rate(df):
    index = 0
    if df['Exchange'] == 'NYSE' or df['Exchange'] ==  'CCSE':
        index += 0
    elif df['Exchange'] == 'AIM' or df['Exchange'] == 'LSE' or df['Exchange'] == 'OFEX':
        index += 1
    elif df['Exchange'] == 'TWSE' or df['Exchange'] == 'TPEX':
        index += 2
    elif df['Exchange'] == 'KOSDAQ':
        index += 3
    elif df['Exchange'] == 'SNSE':
        index += 4
    elif df['Exchange'] == 'BMV':
        index += 5
    elif df['Exchange'] == 'BOVESPA':
        index += 6
    elif df['Exchange'] == 'BVC':
        index += 7
    elif df['Exchange'] == 'BASE':
        index += 8
    elif df['Exchange'] == 'SEHK':
        index += 9
    elif df['Exchange'] == 'ASX':
        index += 10
    elif df['Exchange'] == 'TTSE':
        index += 11
    elif df['Exchange'] == 'JMSE':
        index += 12
    elif df['Exchange'] == 'ENXTAM':
        index += 13
    return exchange_rate.loc[index,float(df['Year'])]

stock_price_adjusted['exchange_rate'] = stock_price_adjusted.apply(get_exchange_rate, axis = 1)

for column in ['Open', 'High', 'Low', 'Close', 'Adj Close']:
    stock_price_adjusted[column] = (stock_price_adjusted[column]/stock_price_adjusted['exchange_rate'])

stock_price_adjusted.shape

stock_price_adjusted.head(1)

# Check how many of the 2006 companies are in financial industry

balance_sheet_df = pd.read_csv('2.18 interpolated balance sheet.csv')
balance_sheet_df.head(1)

temp = balance_sheet_df[balance_sheet_df['isin'].isin(isin_2006)]

def get_num_finance(df):
  return df[['isin', 'is_finance']].drop_duplicates()['is_finance'].sum()

get_num_finance(temp)

# Get quarterly stock price data

q_stock_price = stock_price_adjusted[stock_price_adjusted['ISIN']==isin_2006[0]]

q_stock_price.head(1)

logic = {'Open'  : 'first',
         'High'  : 'max',
         'Low'   : 'min',
         'Close' : 'last',
         'Volume': 'sum'}

q_stock_price = q_stock_price.iloc[:,:9].resample('Q', on='Date').apply(logic)
q_stock_price.reset_index(inplace = True)
q_stock_price = pd.concat([q_stock_price, pd.DataFrame({'ISIN': [isin_2006[0]]*q_stock_price.shape[0]})], axis = 1)

q_stock_price.head()

q_stock_price.shape

for isin in isin_2006[1:]:
  temp = stock_price_adjusted[stock_price_adjusted['ISIN']==isin]
  temp = temp.iloc[:,:9].resample('Q', on='Date').apply(logic)
  temp.reset_index(inplace = True)
  isin_column = pd.DataFrame({'ISIN': [isin]*temp.shape[0]})
  temp = pd.concat([temp, isin_column], axis = 1)
  q_stock_price = pd.concat([q_stock_price, temp], axis = 0)

q_stock_price.shape

q_stock_price.head()

# Export the stock data

#q_stock_price.to_csv('2.18 quarterly stock price.csv', index = False)
