==> Codes to extract data/4. Merge the stock data with the balance sheet data 2.18.ipynb <==
import numpy as np
import pandas as pd

# Load two datasets

stock_price = pd.read_csv('2.18 quarterly stock price.csv')
balance_sheet = pd.read_csv('2.18 interpolated balance sheet.csv')

stock_price.head(1)

balance_sheet.head(1)

print(stock_price.shape)
print(balance_sheet.shape)

# Keep only the companies in stock_price

isin_list = list(stock_price['ISIN'].unique())

len(isin_list)

balance_sheet = balance_sheet[balance_sheet['isin'].isin(isin_list)]

balance_sheet.shape

# We cannot use index to merge them. Hence, try to sort them:

sorted_stock_price = stock_price.sort_values(by=['ISIN', 'Date']).reset_index(drop=True)
sorted_stock_price.head()

sorted_balance_sheet = balance_sheet.sort_values(by=['isin', 'datadate']).reset_index(drop=True)
sorted_balance_sheet.head()

## Check that the order is correct:

temp = sorted_stock_price[['Date', 'ISIN']]
temp.columns = ['datadate', 'isin']

np.sum(temp != sorted_balance_sheet[['datadate', 'isin']])

# Merge the two dataset

temp = sorted_stock_price.drop(columns = ['Date', 'ISIN'])

combined_df = pd.concat([sorted_balance_sheet, temp], axis = 1)

combined_df.head()

## Check that there is no null value:

np.sum(combined_df.isnull(), axis = 0)

# Check the number of firms and number of financial firms in the combined dataset

def check(df):
  num_company = df['isin'].nunique()
  num_rows = df.shape[0]
  print(f'Number of companies: {num_company}')
  print(f'Number of rows: {num_rows}')

def get_num_finance(df):
  num_finance = df[['isin', 'is_finance']].drop_duplicates()['is_finance'].sum()
  print(f'Number of financial firms: {num_finance}')

check(combined_df)

get_num_finance(combined_df)

# Export the combined dataset

#combined_df.to_csv('2.18 combined dataset of stock data and balance sheet data.csv', index = False)

# Try to load the dataset

example_df = pd.read_csv('2.18 combined dataset of stock data and balance sheet data.csv')

example_df.shape
