==> Codes to extract data/5. Eliminating not useful companies.ipynb <==
import numpy as np
import pandas as pd

def check(dataframe):
  num_company = dataframe['isin'].nunique()
  num_rows = dataframe.shape[0]
  print(f'Number of companies: {num_company}')
  print(f'Number of rows: {num_rows}')

def get_num_finance(dataframe):
  return dataframe[['isin', 'is_finance']].drop_duplicates()['is_finance'].sum()

# Import original dataset before elimination

df = pd.read_csv('2.18 combined dataset of stock data and balance sheet data.csv')

# Change a problematic company name

df[df['company_name']=='ADMINISTRADORA DE FONDOS DE']['isin'].unique()

After doing research we notice that they are two completely different companies with the same abbreviation, one is called 'Administradora de Fondos de Pensiones Capital S.A.' and the other is called 'Administradora de Fondos de Pensiones Provida S.A.'. Thus we modify the name:

df['company_name'] = np.where(df['isin']=='CL0000003096', 'ADMINISTRADORA DE FONDOS DE PENSIONES CAPITAL S.A.', df['company_name'])
df['company_name'] = np.where(df['isin']=='CLP7919K1035', 'ADMINISTRADORA DE FONDOS DE PENSIONES PROVIDA S.A.', df['company_name'])

# Check some basic info

check(df)

get_num_finance(df)

df[df['isin']=='CL0000003096']['datadate'].duplicated().sum()

temp = df[df['isin']=='CL0000003096']['datadate']
temp.iloc[0]
temp[temp == '2006-03-31']

# 1. Eliminate based on the 0.9 criteria

df[df['total_assets']/df['total_liabilities'] <= 0.9]

isin_to_be_dropped = list(df[df['total_assets']/df['total_liabilities'] <= 0.9]['isin'].unique())

df = df[~df['isin'].isin(isin_to_be_dropped)]

check(df)

get_num_finance(df)

# 2. Eliminate the companies with correlation equal to -1, 1, or 0

## Functions from model

df['total_assets'].replace(0,0.01,inplace = True)

(df[df['company_name'] == df['company_name'].unique()[0]].shape[0])+1

len(df['company_name'].unique())

### Change fiscal_quarter

series_list = []
for i in range(1,(df[df['company_name'] == df['company_name'].unique()[0]].shape[0])+1):
  series_list.append(i)

temp = pd.DataFrame({'fiscal_quarter': series_list * (len(df['company_name'].unique()))})
df['fiscal_quarter'] = temp['fiscal_quarter']

len(df['company_name'].unique())

temp['fiscal_quarter'].shape

df["datadate"] = pd.to_datetime(df["datadate"])

df.head(5)

temp = df['datadate']
print(type(temp.iloc[0]))

### These are function producing the estimated parameters used in the Zhou(2001) model
def get_companies():
  return df['company_name'].unique()

def get_industries():
  return df['industry_name'].unique()

def get_date():
  return df['datadate'].unique()

def get_num_quarter():
  return len(df['fiscal_quarter'].value_counts())

def get_companies_from_industry(industry):
  return df[df['industry_name'] == industry]['company_name'].unique()

def get_companies_from_country(country):
  return df[df['iso_country_code'] == country]['company_name'].unique()

def get_Vi0(start,company1,company2):
  '''
  This function takes the a start date and two company names to return the starting value of assets
  Output: [v_company1_0,v_company_2_0]
  '''
  v_10 = df[(df['company_name'] == company1)&(df['datadate'] == start)]['total_assets'].item()
  v_20 = df[(df['company_name'] == company2)&(df['datadate'] == start)]['total_assets'].item()
  return v_10,v_20

def get_Ki(start,company1,company2):
  '''
  This function takes the a start date and two company names to return the starting value of Ki := short-term liability + 50% long-term liability
                                                                                                 = total liability -50% long-term liability
  Output: k_company1,k_company_2
  '''
  total_liability_1 = df.loc[(df['company_name'] == company1)&(df['datadate'] == start),['total_liabilities']].iloc[0,0]
  total_liability_2 = df.loc[(df['company_name'] == company2)&(df['datadate'] == start),['total_liabilities']].iloc[0,0]
  # long_term_liability_1 = df.loc[(df['company_name'] == company1)&(df['datadate'] == start),['long_term_liabilities']].iloc[0,0]
  # long_term_liability_2 = df.loc[(df['company_name'] == company2)&(df['datadate'] == start),['long_term_liabilities']].iloc[0,0]
  return  0.75*total_liability_1,0.75*total_liability_2

def get_t(start,end):
  '''
  This function takes the start date and end date of a time horizon and return t
  in Zhou(2001) formula
  Output: t(in years)
  '''
  start_idx = np.where(df['datadate'] == start)[0][0]
  end_idx = np.where(df['datadate'] == end)[0][0]
  return (end_idx - start_idx)/4

def get_sigma_i(company):
  '''
  This function returns the estimated asset volatility of a firm
  '''
  return 0.4

def get_rho(start,end,companies,column,company_1 = None,company_2 = None):
  '''
  This function takes the start date and end date of a time frame and returns the correlation matrix of asset value in a given time horizon
  Note companies is a numpy array containing the name of companies
  Output: A list object:[whole correlation matrix, corr(companyname1,companyname2)] if company_1 and company_2 specified
          the whole correlation matrix otherwise
  '''
  diff_matrix = np.zeros((int(get_t(start,end)*4),len(companies)))
  for i in [i for i in range(0,len(companies))]:
    col_val = np.diff(np.log(df[(df['datadate'] >= start) & (df['datadate'] <= end)&(df['company_name'] == companies[i])][column]))
    diff_matrix [:,i]= col_val
  diff_matrix = pd.DataFrame(diff_matrix)
  corr_df  = diff_matrix.corr()
  if company_1 == None and company_2 == None:
    return corr_df
  else:
    company_1_idx = companies.tolist().index(company_1)
    company_2_idx = companies.tolist().index(company_2)
    output = corr_df[company_1_idx][company_2_idx]
    return output

## Elimination: stage 1

get_rho(get_date()[0],get_date()[3],get_companies(),'Close')

temp = get_rho(get_date()[0],get_date()[3],get_companies(),'Close')
series = np.sum(temp.isnull(), axis = 0)==232
problematic_list = list(series.index[series])

problematic_name = []
for i in problematic_list:
  problematic_name.append(get_companies()[i])
problematic_name

We decide to remove them all, as they have difference in stock price = 0, which makes no sense.

df = df[~df['company_name'].isin(problematic_name)]

check(df)

get_num_finance(df)

## Elimination: stage 2

get_rho(get_date()[0],get_date()[3],get_companies(),'Close')

# Check correlation == 1:
temp = get_rho(get_date()[0],get_date()[3],get_companies(),'Close')
series_1 = np.sum(temp==1, axis = 0)
cov1_list = list(series_1[series_1 != 1].index)
cov1_list

# Check correlation == 0:
series_0 = np.sum(temp==0, axis = 0)
cov0_list = list(series_0[series_0 != 0].index)
cov0_list

# Check correlation == -1:
series_neg1 = np.sum(temp==-1, axis = 0)
covneg1_list = list(series_neg1[series_0 != 0].index)
covneg1_list

problematic_cov_list = cov1_list + cov0_list

problematic_name_2 = []

for i in problematic_cov_list:
  problematic_name_2.append(get_companies()[i])

problematic_name_2

We decide to remove them all.

df = df[~df['company_name'].isin(problematic_name_2)]

check(df)

get_num_finance(df)

# Export the new dataset

df

#df.to_csv('2.22 combined dataset of stock data and balance sheet data.csv', index = False)
